desloppify: extract PermissionManager from AppState god class
Extract ~170 lines of permission checking, requesting, and polling logic
into a focused PermissionManager service. AppState delegates to it via
thin forwarding methods, preserving the existing public API.

- PermissionManager owns permission state, checking, and polling lifecycle
- AppState forwards permission calls for backward compatibility
- PermissionManager.objectWillChange propagated to trigger SwiftUI updates
- Reduces AppState from 732 LOC to 588 LOC
